requirements.txt
# =============================================================================
# TRELLIS-HYBRID-V2 Dependencies
# Pipeline: Image -> Qwen Multi-View -> BiRefNet -> TRELLIS Multi-Image -> PLY
# =============================================================================

# PyTorch and torchvision
--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1

# Numpy (version <2 for compatibility)
numpy<2

# =============================================================================
# Basic dependencies
# =============================================================================
pillow==11.3.0
tqdm==4.67.1
opencv-python-headless
easydict==1.13
scipy==1.16.3
pydantic==2.12.5
pydantic-settings==2.12.0
spconv-cu126==2.3.8

# =============================================================================
# BiRefNet dependencies (required for background removal)
# =============================================================================
kornia
timm

# =============================================================================
# Transformers and AI (4.57.3 for BiRefNet compatibility)
# =============================================================================
transformers==4.57.3
peft==0.18.0
accelerate==1.12.0

# =============================================================================
# FastAPI and web
# =============================================================================
fastapi==0.123.7
uvicorn[standard]
python-multipart
loguru==0.7.3

# =============================================================================
# Utils3d from GitHub
# =============================================================================
git+https://github.com/EasternJournalist/utils3d.git@9a4eb15e4021b67b12c460c7057d642626897ec8

# =============================================================================
# Diffusers from GitHub (for QwenImageEditPlusPipeline)
# =============================================================================
git+https://github.com/huggingface/diffusers

# =============================================================================
# PLY export
# =============================================================================
plyfile>=1.0.0

# =============================================================================
# Model loading
# =============================================================================
safetensors>=0.4.0
huggingface_hub>=0.23.0
modelscope>=1.10.0

# =============================================================================
# Other
# =============================================================================
einops>=0.7.0
ninja
